+++ api/app/models.py
@@ -401,9 +401,7 @@ class Ticket(Base):
 
     threat = relationship("Threat", back_populates="ticket")
     alert = relationship("Alert", uselist=False, back_populates="ticket")
-    ticket_statuses = relationship(
-        "TicketStatus", back_populates="ticket", cascade="all, delete"
-    )
+    ticket_statuses = relationship("TicketStatus", back_populates="ticket", cascade="all, delete")
     current_ticket_status: Mapped["CurrentTicketStatus"] = relationship(
         "CurrentTicketStatus", uselist=False, back_populates="ticket", cascade="all, delete"
     )
@@ -433,6 +431,9 @@ class TicketStatus(Base):
     created_at: Mapped[datetime] = mapped_column(server_default=current_timestamp())
 
     ticket = relationship("Ticket", back_populates="ticket_statuses")
+    current_ticket_status = relationship(
+        "CurrentTicketStatus", uselist=False, back_populates="ticket_status"
+    )
 
 
 class CurrentTicketStatus(Base):
@@ -455,6 +456,7 @@ class CurrentTicketStatus(Base):
     updated_at: Mapped[datetime | None]
 
     ticket = relationship("Ticket", back_populates="current_ticket_status")
+    ticket_status = relationship("TicketStatus", back_populates="current_ticket_status")
 
 
 class Alert(Base):
